# Improve JSON parser allocations