This is a refactoring of the json parser to avoid a maximun of allocations.

bench so far:

```
❯ benchstat before.txt after.txt
name           old time/op    new time/op    delta
JSONParser-16    1.53µs ± 1%    1.63µs ± 4%   +6.38%  (p=0.008 n=5+5)

name           old alloc/op   new alloc/op   delta
JSONParser-16      952B ± 0%      872B ± 0%   -8.40%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
JSONParser-16      12.0 ± 0%       6.0 ± 0%  -50.00%  (p=0.008 n=5+5)
```